# project/docs/swagger.yaml
swagger: '2.0'
...
info:
  version: '1.0.0'
  title: Game Api
  description: This is a game Api created to study API's
schemes:
  - http
# Format of bodies a client can send (Content-Type)
consumes:
  - application/json
# Format of the responses to the client (Accepts)
produces:
  - application/json

################################################################################
#                                    Paths                                     #
################################################################################
paths:
...
  /categories/:
    get:
      summary: get all categories
      tags:
        - Categories
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              categories:
                type: array
                items:
                  $ref: "#definitions/Category"
    post:
      summary: Create a new category
      tags:
        - Categories
      parameters:
      - in: body
        name: category
        description: Category to create.
        schema:
          type: object
          required:
            - name
            - description
          properties:
            name:
              $ref: "#definitions/Category/properties/name"
            description:
              $ref: "#definitions/Category/properties/description"
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
              category:
                $ref: "#definitions/Category"
        400:
          description: Validation error
  /categories/{categoryId}:
    get:
      summary: Get a category by id
      tags:
        - Categories
      parameters:

      - in: path
        name: categoryId
        description: Category id.
        type: integer
        required: true

      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              category:
                $ref: '#definitions/Category'
        404:
          description: Category not found
    put:
      summary: Updates an existing category
      tags:
        - Categories
      parameters:
      - in: path
        name: categoryId
        description: category id to update
        required: true
        type: integer
      - in: body
        name: category
        description:  Category to update.
        schema:
          $ref: '#definitions/Category'
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              category:
                $ref: '#definitions/Category'
        404:
          description: category not found
        400:
          description: validation Error
    delete:
      summary: Delete a category
      tags:
        - Categories
      parameters:

      - in: path
        name: categoryId
        description: Category id to delete
        type: integer
        required: true
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string

        404:
          description: Category not Found
...